Optimize upload slightly
Optimisation to skip conversion of uploaded file if its already a kepub file. Both EPUB and KEPUB files have the same mimetype, so we check the filename instead. This will have some false negatives, in which case the conversion is performed anyway and some performance is lost but the app remains functional.